Add files via upload
I'm uploading the build version of the electronic trading ecosystem for stocks. The output should be a simple log file. Please let me know if you have any questions on the design of the project. THIS IS NOT EASY you need to master the basics to get to this point. Microsoft learn, and google scholar has resources to help but nothing is like doing it yourself.