Add files via upload
Hey this is still a lead up to the clean build that is why I have only the no-clean-build.sh script here. Please at this point you should understand the order and be very familiar with the code base design and implementation. Please at the very least make every headerfile yourself its a great learning opportunity and ya its hard and long and you will have no free time but think of what going to happen when you learn the hard way you can end up in my position. LOL its the only example I got so ya you might want... also so I might have a '*' thats located in a place where you might think its a pointer var but its arithmetic. The Clean build will be checked for the 2 that I know is in but the compiler and the programmer should know just by looking at the code. I'm just putting it out so when your writing this code base you know and can look up at this point.