Add files via upload
BUILD VERSION 1.0. 
README: The reason that this upload is the first that is meant to be build is due to several points that every Computer Scientist should follow. First I wanted you to understand the problem and know which is creating an electronic trading platform for the finance sector.  Understanding finance is hard and complex even when starting from a strong math back round. Second this was focusing on learning low level programing for better cloud computing. Just google "cloud computing '/' AI". Important, I typed each page by had, Yes no copying and pasting. Purly did it to show it i possible. Albeit, long, hard, and you will burn out personally I got a lot going on with trying career wise and trust me when I say "I could of been doing something that will help me get to the place I want" but wanted to show you can do it and you will get better its just HARD. Last thing tech and the finance like so many other important industries have given birth to "FINTECH" and I wanted to get across that this is VERY important and much more research needs to be done in this field. "Fin-Tech" leads into data scientist and other "BUZZ" words that are taking off but this subject is far more important and will govern the future of money flow with or without streaming-blocks of digital money "cryptocurrency". I will be uploading more and filling extra content to help people that need the help. You can contact me if you got my number or on social media. kevin heart